+++ lib/ready_set_go_web/controllers/athlete_html.ex
@@ -1,0 +1,13 @@
+defmodule ReadySetGoWeb.AthleteHTML do
+  use ReadySetGoWeb, :html
+
+  embed_templates "athlete_html/*"
+
+  @doc """
+  Renders a athlete form.
+  """
+  attr :changeset, Ecto.Changeset, required: true
+  attr :action, :string, required: true
+
+  def athlete_form(assigns)
+end